+++ ACadSharp/Tables/Collections/LayoutsTable.cs
@@ -1,8 +1,10 @@
 ﻿using ACadSharp.IO.Templates;
 using ACadSharp.Objects;
+using System;
 
 namespace ACadSharp.Tables.Collections
 {
+	[Obsolete("This object should be a collection of layouts")]
 	public class LayoutsTable : Table<Layout>
 	{
 		public override ObjectType ObjectType => ObjectType.INVALID;
